[FV] FV Speed Up #5: survival times are stored in the order they are observed, sorted at the end
The 3 unit tests on the FV estimation in test_estimators.py pass.

The execution time of the FV estimation process using test_QB.py with K=40, N=400 particles, T=80000, J=20, does NOT change significantly. In fact, the number of survival times is small (~ 30), thus the increased processing speed is not observed.

The only function changed is run_simulation_mc() in simulators.py.