[FV] FV Speed Up #6: reactivation particle is selected using Robins algorithm to avoid random number generation
Note that the Robins algorithm is implemented but it is NOT run in the current implementation because I am not really sure that I want to use it... as it may have an impact in the estimated probability by FV...(?)
This is the reason why the 3 unit tests on the FV estimation in test_estimators.py still pass. If I decide to use the Robins algorithm, I would need to change the results of these 3 unit tests.

In terms of performance, in an initial test using test_QB.py on 2 replications with K=40, N=400 particles, T=80000, J=20, the execution time of the FV-only part goes down from 14 min to 10 min, but only in one of the two replications; in the other replication, the execution time stays at the same 10 min time spent in the version where the reactivated particle is randomly selected.

The affected methods by this change are:
- reactivate_particle()
- choose_particle()