[FVRL] Fixed the clipping version of the algorithm
The clipping was not implemented correctly because the delta(theta) value was clipped between +1 and +1, instead of between -1 and +1!

Note that this bug did NOT affect the results presented in the ICML-2022 nor NEURIPS-2022 paper because it was introduced AFTER I had ran the FVRL simulation in Jan-2022.

I also fixed the parsing of the `clipping` parameter when running the simulation through simulators.py which had the conversion to boolean of the input character value when running the program from the command line missing (e.g. the conversion from "True" (str) to `True` (bool)).