[utils] measure_exec_time() now reports the CPU time as well
Goal: Have more control on the execution time by making it independent of other processes running in the computer when the program is executed.

However, I haven't seen much difference between the CPU time and the total execution time in the few tests I have done so far (about 20-30 executions).