[REFACTOR] Moved classes and functions from simulators.py to simulators/ directory and created run script of FVRL learner
All unit tests pass and other tests run fine. The following were tested:
- test_estimators.py (both "value_functions" and "fv")
- test_MC_lambda.py
- test_TD_lambda.py
- test_QB.py (not run in test mode but in script mode)
- test_simulators.py (new file!)

Note: these tests can be run from a CMD window from within the test/ directory of the project, using the following command:
python -m pytest
Note however that ALL unit tests found in `test*` files are run, regardless of the test suite we define in the main section. It's as if we ran the unit tests using `unittest.main()`.
Note also that 2 out of 4 tests in test_QB.py fail (F) and 1 gives error (E), but this is ok because right now test_QB.py is NOT intended to be run in unit test mode. Also, the two tests in test_simulators.py on the SimulatorQueue class using the IGA learning strategy for theta fail, as the FVRL process is not prepared for IGA learning.

General: I reorganized where classes and enums are defined.
For enums, I moved enums previously defined in agents/queues.py to agents/learners/__init__.py (for learner-related enums), and to agents/policies/__init__.py (for policy-related enums).
Note that the PolicyTypes enum defined in policies/__init__.py currently holds only policies that are meaningful for queues (in fact, this enum was previously defined inside agents/queues.py), but in the future they could include policies that are meaningful for other environments as well.

DETAILS:
- agents:
  - __init__.py: the GenericAgent class now defines an act() method which is "Not implemented", in order to indicate that this method should be defined by classes inheriting from this class.

- simulators/: new files:
  - __init__.py: functions normally used in simulations are defined here (e.g. step(), check_dones(), etc.).

  - discrete.py: the class Simulator is defined here, which is used to run experiments on the TD(lambda) methodology to learn the state value function V on gridworlds and mountain car environments. This class was taken from the no longer existing simulators.py file.
  Since the class is now defined inside discrete.py the TODO message stating that the class should be renamed to SimulatorDiscreteEpisodic was removed. Furthermore, also the "Episodic" keyword in the proposed new name was also erased as I think the "Episodic" characteristic is no longer relevant as the simulate() and run() methods defined therein have a parameter called `max_time_steps` which can be used to stop an "episode" after a certain number of steps have been run --as is the case in the SimulatorQueue class defined in simulators/queues.py, which simulates non-episodic tasks.

  - queues.py: the class SimulatorQueue is defined here, which is used to run experiments on the Flemint-viot estimation of the blocking probability and on the FVRL learner of the optimum theta defining the acceptance policy of new jobs arriving to a queue system.
  I also placed here the functions that are used to test these learning methods, which were previously defined in simulators.py.

- utils/computing.py: the following functions were moved to the brand new simulators/queues.py file as they are very specific to the simulation of queues via Fleming-Viot:
  - compute_nparticles_and_nsteps_for_fv_process()
  - compute_rel_errors_for_fv_process()

- test/test_QB.py: changed parameter T from 200*N to 50*N, because I find 200*N a too large value for the number of events used in the simulation of the single particle that is used to estimate P(T>t) and E(T).

- New file test/test_simulators.py which has been created to hold the unit tests previously present in simulators.py on the Simulator class and on the SimulatorQueue class defined in the newly created simulators/ directory.

- New file run_FVRL.py which can be used to run the FVRL learner of the parameterized acceptance policy in a single-server queue system.