[REFACTOR] Reorganized tests and executions previously run by test_QB.py; new script run_FV.py
Goal: Separate unit tests from run scripts as they should not mix up.

General: Fixed imports of the compute_job_rates_by_server() function which was incorrectly (indirectly) imported from Python.lib.simulators.queues instead of importing it directly from where it is defined, namely in Python.lib.utils.computing (this indirect import generated some errors and that's how I found it).

All 25 unit tests pass (note that previously I mentioned *32* unit tests, and that's because I was counting the tests defined in test_script4yao.py which currently is not part of the repository), and the execution of run_FV.py works both in "N" and "J" modes, as defined by script parameter `analysis_type`. The executions in both modes "N" and "J" in run_FV.py give the same results as their executions in the version before refactoring, which was checked by comparing in TextPad the output generated in the Python console by the old and the new implementation.

DETAILS:
- New script run_FV.py that is used to run the FV estimation of the average reward, thought of as blocking probability --i.e. the reward for blocking is 1.

- test_QB.py:
  - The test_fv_implementation() and the analyze_estimates() methods originally defined in the Test_QB_Particles class were converted to functions and renamed respectively to:
     - analyze_convergence() --> as it analyzes the convergence of the estimator as N increases.
     - analyze_absorption_size() --> as it analyzes the effect of the absorption size J in the FV estimator quality.
  Both functions were moved to the new script run_FV.py, as they are part of *running* the FV estimation method.
  - The plot_results_fv_mc() function was moved to the new script run_FV.py.
  - The Process enum was moved to the new script run_FV.py.
NOTE: This file will probably disappear or used to define relevant unit tests of the FV process applied to queue blocking, as currently all tests and functions are either not used, irrelevant, or old (named with prefixes "deprecated_" or "no_").

- simulators/__init__.py: This file now holds the function previously defined in run_FVRL.py called generate_queue_environment_and_agent() as it is now called also by run_FV.py. The function is used to setup the queue environment and agent that will interact with it during the simulation carried out to apply FV, both in the estimation and in the control problem.

- queues/__init__.py & environments/queues.py: New method setCapacity that sets the capacity K of the queue. These are used in the analyze_absorption_size() function to run the FV estimation process on the same queue but with different capacities that are set in a loop.

- parameterized.py: In both classes receiving a real number as parameter theta (as opposed to a list), the assertion that theta < capacity is now only executed when capacity < Inf, as, when Inf, theta is also allowed to be Inf. This has an impact in run_FV.py where the parameterized policy is initialized with capacity = Inf in analyze_absorption_size() function because the capacity of the queue in the analysis performed may be more than one and is set inside the loop.

Other less relevant changes:
- simulators/discrete.py: fixed a call to the setInitialDistribution() method which was incorrectly referenced.

- simulators/queues.py: removed unnecessary semicolons and continuing command indicator (`\`) from a few lines. These were pinpointed by the "Problems" tab in PyCharm.

- run_FVRL.py: the object `env_queue_mm` was renamed to `env_queue` just to simplify the name.