[FVRL] Values of N and T are updated at each learning step
Goal: Reduce number of computations for small values of the blocking size K.

Notes:
1) The update of N and T performed are in detriment of performance when the initial values are set in a scenario where the blocking size K is larger than the optimum as in that case the values of N and T get *reduced* from their original values as learning occurs (as the new blocking size will get smaller).
On the contrary, if we start with a small K and then its value is increased through learning, the performance will be better than keeping the original N and T values.
However, the most common case when showcasing the FVRL method is starting at K larger than the optimum... therefore the former situation applies.

2) In a true scenario where we don't know the actual relationship between N, T and the respective errors of estimation of Phi and E(T_A), we can still update them using some heuristics that recommends an increase of N as J/K decreases, and a increase of T as J increases.