[FVRL] The number of events per learning step in MC learning is now constant
The number of events per learning step in Monte-Carlo learning is now set to the total number of events in FVRL divided by the number of learning steps (instead of being set to the same number of events observed by FVRL at the *respective* learning step).
The goal is that MC learning runs the queue for the same number of steps per learning step, which better reflects the situation in which a Monte-Carlo simulation would be run (which has no knowledge of the existence of FVRL).