[FV] run_FV.py, plot_results_fv_mc(): new parameter `plot_violin_only` and bug fix when grouping values in MC are not unique
Goal of new parameter: Generate only the violin plots when comparing the FV vs. MC results so that we don't get overwhelmed with all the 4-5 generated plots.

The bug fix has to do with the situation when the grouping values plotted on the X-axis of the MC results are not unique. This should be rare but it has to do with the fact that the grouping variable defining the X-axis values of the MC plot is NOT the same as the grouping variable defining the X-axis values of the FV plot (e.g. number of particles): in fact the grouping variable for the MC plot is always the number of cycles observed in the MC simulation for each grouping value on the X-axis of the FV plot, and thus its values could repeat if the same number of cycles was observed in MC simulations for two different e.g. number of particles used in the FV simulation. At first glance, we would think this should not happen or it would happen with very low probability, but it happened already... (I don't recall the settings).