[FV / FVRL] New scripts plot_FV.py and plot_FVRL.py used to generate plots for saved results
Goal: Generate the plots showing the results of the simulations I left running at the IRIT cluster in November / December 2022, which are many because I wanted to investigate the impact of the different hyperparameters involved in the FV and FVRL estimators.

Note: For a correct import of functions defined in run_FV.py (e.g. plot_results_fv_mc()), the code run by executing the script was included inside an `if __name__ == "__main__"` block.