[FVRL] Implemented the possibility of setting the learning rate alpha proportional to the inverse of the min. component of the gradient
This is implemented via an ad-hoc approach, i.e. we need to uncomment certain lines in the LeaPolicyGradient.learn_linear_theoretical_from_estimated_values() method in policies.py in order to use this strategy.

IMPORTANT: When activated, this choice of alpha is done at EVERY learning step, which may not be the best option as it introduces quite a bit of noise in the learning process... Urtzi suggested setting JUST the INITIAL alpha with this strategy, and using the REAL stationary probabilities to compute the gradient (together with the estimation of Qdiff of course); o.w. if we use the estimated stationary probabilities to set the initial alpha, the estimated probability associated to the class with the smallest load rho will most likely be 0 and thus will NOT be taken into account when computing alpha... which will be thus strongly underestimated. Once we set the initial alpha, we reduce it as usual.
However, this strategy is currently NOT implemented, and I have dropped it for now because I found another way of defining the loss network system on which interesting learning takes place and on which we can show the effectiveness of Fleming-Viot over Monte-Carlo.