[FVRL] Found a reasonable set of system parameters for a 2-class loss network that allows showing interesting learning in reasonable time
This system parameters allows us to:
a) Have an optimal blocking that is NOT trivial (e.g. [3, 6] instead of [6, 6] when K=6, which is kind of trivial, as it says "block as little as possible").
b) Be able to learn rather fast (e.g. < 50 learning steps).
c) That the blocking probability is small enough (e.g. ~ 1E-5) in order to show the advantage of Fleming-Viot over Monte-Carlo.

On a 2-class loss network the best parameters that were found so far with optimum Ks = [4, 6] is:
- K = 6 (system's capacity)
- rhos = [0.5, 0.3]
- lambdas = [1, 5]
- costs = [2000, 20000]
The blocking probability of each class are of the order of 1E-5 so we expect that we are able to show the advantages of Fleming-Viot over Monte-Carlo, perhaps by not choosing too large N and T values (to avoid that Monte-Carlo learns) (e.g. N=70, T=50 with J somewhere between 0.1 and 0.3).

The set of system parameters were found by imposing the condition that we want learning steps of theta that are of the order of delta(theta) ~ 1 for ALL theta dimensions. This is achieved by making the contributions to the gradient of the average state value v (coming from each theta dimension) be of comparable scale. At the same time we want to set blocking costs that are not toooo different among classes because it is not so realistic (ref: Urtzi).

Note that we leverage the existence of two degrees of freedom in setting system parameters:
- p(x) ~ rho^K (i.e. is a function of the rhos)
- arrival probability ~ lambdas (i.e. NOT rhos)

So, we start by setting a common rho for all classes (say 0.5), then we set increasing lambdas (so that we have some difference among the classes (e.g. [1, 5]). These two quantities define the service rates mu as lambdas/rhos.
Considering that a homogeneous rho implies that p(x) is about the same for all classes, we when choose costs for the different job classes so that their contribution to the expected cost (i.e. a sum of terms of the form p(x can block i) * cost(blocking(i)) * lambda(i) / Lambda) is about the same. This gives us a nominal set of costs that we then tune a little bit so that the contribution to the expected cost from each class is not exactly the same, o.w. the optimum theta will be trivial (say [6, 6] when the system's capacity K = 6).

For instance for a 3-class system with rho = 0.8 for all classes and lambdas = [1, 5, 8], we get mus = [1.25, 6.25, 10.0] and costs = [42, 8, 5] which we then set to [100, 10, 1] in order to avoid the trivial optimum theta.

These details are also explained in run_FVRL.py under a commented section titled "BLOCKING COSTS" just prior to the definition of the blocking costs.