[FVRL] Removed non-valid "boundary" states as start states for FV in multidimensional systems
SUMMARY: States that are NOT at the boundary of the complement set Ac are no longer selected. Essentially, states that have more than one dimension at a dimension boundary are NOT allowed (e.g. [2, 2, 2] when J = 2 for all dimensions).
This fix has 2 effects:
(i) it removes the overestimation of the stationary probabilities of blocking (because P(T>t) is normally overestimated)
(ii) it increases the particles' absorption rate, thus making them more likely to be closer to the blocking states of interest, hence a signal is received from more blocking states --and this has a direct impact on learning via FVRL!

DETAILS:
- simulators/__init__.py: SetOfStates class: new methods getStateBoundaries() and states_are_given_explicitly() that were created to express purpose of what we are doing when calling these methods.

- simulators/queues.py:
  - run_simulation_fv(): the start state is now correctly selected in the sense that only states that are ACTUALLY at the boundary of the Ac set are selected. For instance, states such as (2, 2, 2) are NOT in the boundary when J = [2, 2, 2] because it is NOT possible to transition from there to the absorption set in just ONE step (=> the probability of leaving Ac from there in one step is zero, making that state not part of the boundary of Ac --by definition of boundary).

  - Fixed debugging and plotting in a few places (look for DEBUG_ESTIMATORS and DEBUG_TRAJECTORIES and `if plot` conditions).