[FV] Made estimation of Phi a little more efficient
SUMMARY: In update_phi() inside run_simulation_fv() defined in simulators/queues.py (which is called every time there is a step in the Markov chain being simulated) only the previous and current visited states are checked for value update in the `dict_phi` dictionary. In fact, these are the only ones where Phi(x) could change at every step.

Note: The impact of this change is not really noticeable in small dimension problems (e.g. 2D-states), but if we attack problems with larger dimensions I think it should be noticeable, because the set of blocking states may become quite large in that case.