New convert_str_argument_to_list_of_type() function used by unit tests; removed import of AcceptPolicyType
Goal: Fix incomplete commits that were previously done that make unit tests pass when cloning the remote repository.
Note that the AcceptPolicyType is a new enum class that is expected to be implemented when dealing with job acceptance policies other than threshold-type (e.g. trunk-reservation policies).

One other small thing was fixed, namely the calculation of Ks in the SimulatorQueue.run() method of simulators/queues.py so that its value is scalar in 1D problems and a list in multidimensional problems.