[FVRL-Bug] Fixed the calculation of the set of multidimensional blocking states (e.g. Loss Network)
The states (x(1), x(2), ..., x(L)) whose x(i) values are all smaller than the respective blocking size (x(i) < K(i) for all i) but whose sum is equal to the system's capacity were NOT previously included in the set of blocking states.

This is crucial for the correct estimation of the expected blocking cost!