[FV] Fixed computation of required N / relative phi error by using the correct value (K-J+1)/2 instead of (K-J)/2
This fix just impacts the sizing of the FV system in terms of the number of particles to use. After the fix, the number of required particles to achieve a given expected relative error in Phi is larger than before the fix (because the blocking probability is proportional to rho^[(K-J+1)/2] which is smaller than rho^[(K-J)/2] assumed before).

*** NOTE: No simulations were run yet using this new definition of the computation of N, but things should behave pretty similarly to before the change. ***