[Gridworlds] Changed the terminal_rewards parameter into a dictionary of rewards indexed by state
Goal: Make the terminal rewards setting both more intuitive and more robust to bugs (as until now the rewards should be given in increasing order of the state 1D index since the states are passed into parameter terminal_states as a SET (which stores the values in increasing order).

In addition, the number of terminal rewards NO LONGER has to coincide with the number of terminal states: if a terminal state is NOT present in the terminal rewards dictionary, its reward is set to 0 (and a warning message is shown when retrieved), and if non-terminal state is present in the terminal rewards dictionary, the reward is ignored (and no warning is shown).

This change is done in class EnvironmentDiscrete of environments/__init__.py.