[Learners] Fixed the inconsistency between self.rewards (in GenericLearner) and self._rewards (in Learner)
Before this fix, the length of self.rewards (the history of rewards stored/updated ONCE, at the end of the episode) was one element shorter than self._rewards (the history of rewards udpated DURING the episode, as rewards are observed).

The fix was done by initializing the self.rewards attribute in GenericLearner to `[0]` instead of the empty list `[]`.
Also, in discrete/mc.py, the use of self._rewards was replaced with self.rewards in an attempt to be closer to extending the Monte-Carlo learning to the average reward criterion case.

On another matter, parameter reset_time in was renamed to reset_learning_epoch for increased clarity of what is being reset, namely the index (self._t) that is increased (or should be increased) every time learning occurs.

All 40 unit tests pass.