[Renewal] Implemented the estimation of the stationary distribution using renewal theory
All 16 unit tests in the affected unit test file test_estimators_discretetime.py pass.

On an aside, a new getter method was defined in EnvironmentDiscrete class to retrieve all the states of the environment without the need to know the name of the attribute holding that information: getAllStates()

DETAILS:
- simulators/discrete.py:
  - run(): New parameter set_cycle that defines ALL the states in the set whose entrance defines a cycle. The method returns the number of cycles observed, the state counts and the expected cycle time as part of the `learning_info` dictionary.

- test_estimators_discretetime.py: new tests on the labyrinth with obstacles running both the MC learner and the TD(lambda) learner that estimate the stationary probability both from the classical point of view (i.e. by visit frequency) and from the renewal theory point of view (i.e. by leveraging cycles).
Note that these estimations are independent of the learning methodology, be it via Monte-Carlo or TD(lambda) and be it under the discounted average reward criterion. These tests were implemented in class Test_EstDifferentialStateValueV_EnvGridworldsWithObstacles.