[Learners] The Learner class honours the value of parameter reset_value_function when resetting
In the learners.episodic.discrete.Learner class the reset() method now calls the super class reset() using the value of `reset_value_functions` passed as parameter, so that we can prevent the value functions from being reset at the beginning of the episode, if so is wished. This could be useful if we want to continue learning the value functions from an initial learning process.