[Learners] Implemented a first version of the action value function (Q) learning in discrete-time learners
NOTE: I wrote "first version" because I realized that the currently implemented learner of Q(s,a) is NOT the correct way of learning the action value function. In fact, I am using the wrong `delta` value, namely the delta value of the *state* value function V(s). The correct implementation involves the delta value of the action value function Q(s,a) which depends ALSO on the next action taken. The latter depends on the strategy followed, whether e.g. SARSA, Q-learning or Expected SARSA. I think the best option is to use Expected SARSA so that we do not prioritize any future action for the bootstrap estimate of Q(S(t), A(t)).
Also, we would need to correctly update the learning rate value according to, not only the state visit frequency, but also according to the *action* visit frequency, which currently is not taken into account in the discrete.Learner class (which just stores the different alphas on the different states, not on the different state and actions).

DETAILS:
- environments.py:
  - EnvironmentDiscrete class: new attributes `state_space` and `action_space` that store the state space and the action space of the environment as spaces.Discrete objects in an attempt to make use of the capabilities of the gym package in handling environment characteristics.
HOWEVER, note that the Discrete object is NOT iterable!! (so we cannot do `for s in self.state_space:` as I expected... we still need to do `for s in range(self.state_space.n):`, etc., which is not so convenient (as I already had the retrieval of the number of states and the number of actions in place). In addition, the state and action values are actually indices (i.e. the same strategy I am already using!). Nevertheless, the object has two convenient methods: contains() to check whether an element belongs to the space, and sample() to sample a value from the space (optionally setting a seed with Discrete.seed()).

- simulators/discrete.py: all the methods that return the state values (e.g. _run*()) now also return the *action* values. Note that the returned objects here are array-like objects (containing the values), NOT more complex objects such as StateValueFunctionApprox or ActionValueFunctionApprox which would allow us to perform more complex retrievals of the value functions information (e.g. in the action value function case, we could use the getValue(state, action) method to retrieve the Q value for a state and action without worrying where in the array of Q values is the value of the state-action stored! (which requires knowledge of the implementation of the ActionValueFunctionApprox class...).

- learners/episodic/discrete/mc.py: implemented the Q attribute as an ActionValueFunctionApprox attribute (before the attribute was set to None), which implied defining a new method called updateQ(). Take a note however of the NOTE written above, about the correct and incorrect learning of the Q values.

- learners/value_functions.py:
  - The method setWeight() was renamed to _setWeight() because the method is only valid for the dummy features case... so we want to make the method "private".
  - The storage of the action values in ActionValueFunctionApprox was changed so that values are grouped by state in the X feature matrix, which I find more intuitive than the previous storage where values were grouped by actions (because states usually come before actions).

- test_estimators_discretetime.py & test_TD_lambda.py: updated all the calls to the run() method by adding one more output to the tuple of returned elements by the method, following the change on the returned values by the methods responsible for running the simulation in the discrete.Simulator class.