[FVAC] First implementation of the FVAC algorithm: Fleming-Viot Actor Critic policy learning
This first implementation consists of updating the policy parameter (using the neural network model defined in the PolNN class and leveraged by the LeaActorCriticNN class) using the estimates of the value functions V(s) and Q(s,a) obtained by an FV simulation of a labyrinth with obstacles environment. A new method in the LeaActorCriticNN class called learn_from_estimated_value_functions() sweeps over all possible states and actions whose advantage value and log-policy contribute to the loss function that is minimized at each learning step.
The policy converges to the optimum policy on a 3x4 labyrinth with one obstacle on 100 learning steps, both using the FV learner and the TD(0) learner of the value functions under the average reward criterion.

However, this policy learning process seems not to be as effective as learning online, i.e. by simulating trajectories on the current policy and contributing to the loss function from each state and action visited by those trajectories, where learning seems to happen in a fewer number of learning steps.

The process is tested at the first entry in tests.py which was taken from its original location at the locally stored scratch-rl.py file (i.e. a file that is stored in the PyCharm installation directory, and thus not tracked by this repository).