[FVAC] The advantage function now uses the average Q over all actions instead of V as reference value
This change makes the loss function DECREASE (instead of increasing, as was the case before!) while the neural network model learns the optimum policy.
The reason is that the sign of the advantage function is now the correct one... before this change, the reference value used for the advantage function (V(s)) didn't always fall in between the Q(s,a) values for the different actions, and thus the reference value was not actually functioning as a proper baseline.