[ActorCritic] Implemented the actual Actor-Critic policy learning, i.e. ONLINE policy learning with CRITIC
The "ONLINE" part means that the policy can now be learned online, also when we learn the value functions separately from the policy.

The "CRITIC" part means that the advantage function is NOT learned at the same time as the policy but is instead provided as an input, previously learned (in such case, the Q(s,a) values are passed to the learn() method).

In terms of results, using a critic reduces the variability in the value functions learning process (this can be seen from the plots of the V(s) and Q(s,a) values estimated at each policy learning step and generated by tests.py. This impacts the variability in the policy learning process, which is also reduced.

DETAILS:
- policies.py: new parameter `action_values` in the learn() method that optionally receives the Q(s,a) values of the action value function previously learned.
This change implied that the average reward is now computed by the method itself (in terms of the rewards observed during the policy learning) instead of being retrieved, as before, from the value function learners (which previously learned the value functions at the same time as the policy).

- test_optimizers_discretetime.py:
  - setUpClass():
    - Adaptive TD(lambda) learner for the value functions added.
    - New parameter `learning_criterion` that defines the reward criterion used to learn the value functions, either AVERAGE or DISCOUNTED.

- tests.py:
  - New possibility of value function learner: adaptive TD(lambda).
  - New possibility of policy learning for the Actor-Critic case (i.e. where value functions are learned separately from the policy), implemented via the new `policy_learning_mode` execution parameter, which can take the values "online" or "offline".