[FV] Fixed the bug where I failed to reset the discrete-time FV learner at the start of the FV simulation
WARNING: This is a SELECTED commit, i.e. something could fail because I may have forgotten to commit a change.

The fix is done in the _run_simulation_fv() method of the discrete.Simulator class.

The expected results of the affected test in test_estimators_discretetime.py (test_Env_PolRandomWalk_MetFV) was updated.

This fixes the bug where the value of the *learning rate alpha* for each state was already smaller than 1 at the start of the FV simulation process (--of course, only when adjust_alpha=True). In fact, the initial alpha value at the start of the FV simulation came from the final alpha values obtained at the end of the single Markov chain simulation that is used to learn the expected reabsorption time E(T_A).
We don't want this to happen because it implies that learning from FV would be slowed down by a smaller learning rate, which may not be sensible when learning from RARE events, since the single Markov chain simulation (defining the initial alpha values) may not have visited the rare events enough to learn value functions of the visited states (which had their alpha value adjusted) at a reasonable accuracy.

Note that, in this respect, the FV estimation of value functions in a 2D labyrinth now gives an estimate that is much closer to the TD estimate than was the case before this change (see the Test_EstDifferentialStateValueV_EnvGridworldsWithObstacles.test_Env_PolRandomWalk_MetFV unit test in test_estimators_discretetime.py).

In addition to the above and in order to be consistent with what is done in the REGULAR Markov chain simulation performed by discrete.Simulator._run_single(), the simulation time counter for the FV simulation is now initialized at -1 (as opposed to 0). This means that the first transition of the FV particle system now happens from time t=0 to time t=1 (instead of from time t=1 to t=2).

Finally, I removed the self.reset() method in the discrete.Simulator class because this was simply calling the learner's reset() method. So, now the learner's reset() method is called instead.

I also improved the comments here and there, and more importantly a couple of variable names in LeaFV (used for the iterative update of the FV integral when survival times are observed in increasing order), and in functions related to the estimation of Phi in estimators/fv.py (related to e.g. renaming state_or_buffer_size --> state or adding the possibility that states_of_interest is None in the initialize_phi() function).

All 49 unit tests pass.

----------------------
Collected 49 items
test_FV.py .                                                             [  2%]
test_QB.py .                                                             [  4%]
test_TD_lambda.py .........                                              [ 22%]
test_environments_gridworlds.py ....                                     [ 30%]
test_estimators_continuoustime.py ....                                   [ 38%]
test_estimators_discretetime.py ...................                      [ 77%]
test_optimizers_continuoustime.py .....                                  [ 87%]
test_simulators.py .                                                     [ 89%]
test_simulators_queues.py .....                                          [100%]